migrate: move vetCenter to products directory
- Migrated vetCenter product files to src/products/vetCenter/
- Moved and renamed index.tsx -> template.tsx
- Moved and renamed index.test.tsx -> template.test.tsx
- Moved query.ts from src/data/queries/ to src/products/vetCenter/
- Moved test file and renamed vetCenter.test.tsx -> query.test.ts
- Moved formatted type from src/types/formatted/ to src/products/vetCenter/formatted-type.ts
- Moved mock file from src/mocks/ to src/products/vetCenter/mock.js
- Updated all import statements across the codebase
- Fixed import paths in migrated files
- All TypeScript errors resolved
- All tests passing (14 query tests, 12 template tests)
- Updated migration guide with completed status
- Added CODEOWNERS entry for new product directory
- Verified dev server runs and vetCenter pages load correctly
