+++ src/data/queries/index.ts
@@ -47,7 +47,7 @@ import * as Table from '@/data/queries/table'
 import * as VamcEhr from '@/data/queries/vamcEhr'
 import * as VamcSystem from '@/data/queries/vamcSystem'
 import * as VamcSystemVaPolice from '@/products/vamcSystemVaPolice/query'
-import * as VetCenter from '@/data/queries/vetCenter'
+import * as VetCenter from '@/products/vetCenter/query'
 import * as VetCenterHealthServices from '@/data/queries/vetCenterHealthServices'
 import * as VetCenterOutstation from '@/data/queries/vetCenterOutstation'
 import * as VetCenterLocationListing from '@/data/queries/vetCenterLocationListing'
