+++ src/pages/[[...slug]].tsx
@@ -51,7 +51,7 @@ import { PressReleaseListing as FormattedPressReleaseListing } from '@/products/
 import { ResourcesSupport as FormattedResourcesSupport } from '@/types/formatted/resourcesSupport'
 import { StaffProfile as FormattedStaffProfile } from '@/products/staffProfile/formatted-type'
 import { StoryListing as FormattedStoryListing } from '@/products/storyListing/formatted-type'
-import { VetCenter as FormattedVetCenter } from '@/types/formatted/vetCenter'
+import { VetCenter as FormattedVetCenter } from '@/products/vetCenter/formatted-type'
 import { HealthCareLocalFacility as FormattedHealthCareLocalFacility } from '@/types/formatted/healthCareLocalFacility'
 import { VamcSystem as FormattedVamcSystem } from '@/types/formatted/vamcSystem'
 import { VamcSystemVaPolice as FormattedVamcSystemVaPolice } from '@/products/vamcSystemVaPolice/formatted-type'
@@ -73,7 +73,7 @@ import { PreviewCrumb } from '@/templates/common/preview'
 import { ResourcesSupport } from '@/templates/layouts/resourcesSupport'
 import { StaffProfile } from '@/products/staffProfile/template'
 import { StoryListing } from '@/products/storyListing/template'
-import { VetCenter } from '@/templates/layouts/vetCenter'
+import { VetCenter } from '@/products/vetCenter/template'
 import { Wrapper } from '@/templates/layouts/wrapper'
 import { HealthCareLocalFacility } from '@/templates/layouts/healthCareLocalFacility'
 import { DoNotPublishError } from '@/lib/drupal/query'
